java tests
