+++ java/test/behaviour/query/language/insert/BUILD
@@ -26,17 +26,18 @@ load("//java/test/behaviour:rules.bzl", "typedb_behaviour_java_test")
 typedb_behaviour_java_test(
     name = "test",
     srcs = [
-        "GetTest.java",
+        "InsertTest.java",
     ],
-    test_class = "com.vaticle.typedb.driver.test.behaviour.typeql.language.get.GetTest",
+    test_class = "query",
     data = [
-        "@vaticle_typedb_behaviour//typeql/language:get.feature",
+        "@vaticle_typedb_behaviour//query/language:insert.feature",
     ],
     connection_steps_core = "//java/test/behaviour/connection:steps-core",
     connection_steps_enterprise = "//java/test/behaviour/connection:steps-enterprise",
     steps = [
-        "//java/test/behaviour/typeql:steps",
         "//java/test/behaviour/connection/session:steps",
+        "//java/test/behaviour/query:steps",
+        "//java/test/behaviour/util:steps",
     ],
     deps = [
         # Internal Package Dependencies
@@ -48,8 +49,8 @@ typedb_behaviour_java_test(
     runtime_deps = [
         "//java/test/behaviour/config:parameters",
     ],
-    size = "large",
-    visibility = ["//visibility:public"],
+    size = "enormous",
+    visibility = ["//visibility:public"]
 )
 
 checkstyle_test(
