+++ java/test/behaviour/query/language/insert/InsertTest.java
@@ -19,7 +19,7 @@
  * under the License.
  */
 
-package com.vaticle.typedb.driver.test.behaviour.typeql.language.update;
+package com.vaticle.typedb.driver.test.behaviour.query.language.insert;
 
 import com.vaticle.typedb.driver.test.behaviour.BehaviourTest;
 import io.cucumber.junit.Cucumber;
@@ -31,10 +31,10 @@ import org.junit.runner.RunWith;
         strict = true,
         plugin = "pretty",
         glue = "com.vaticle.typedb.driver.test.behaviour",
-        features = "external/vaticle_typedb_behaviour/typeql/language/update.feature",
+        features = "external/vaticle_typedb_behaviour/query/language/insert.feature",
         tags = "not @ignore and not @ignore-typedb-driver-java"
 )
-public class UpdateTest extends BehaviourTest {
+public class InsertTest extends BehaviourTest {
     // ATTENTION:
     // When you click RUN from within this class through Intellij IDE, it will fail.
     // You can fix it by doing:
